refactor: Heavily refactored the edit component in order to split it into an insert and an update component
