+++ crates/bld_ui/src/app.rs
@@ -1,7 +1,7 @@
 use crate::pages::{
     home::{
-        CronJobs, CronJobsEdit, Dashboard, History, Home, Monit, PipelineInfo, Pipelines,
-        RunPipeline,
+        CronJobInsert, CronJobUpdate, CronJobs, Dashboard, History, Home, Monit, PipelineInfo,
+        Pipelines, RunPipeline,
     },
     login::Login,
     not_found::NotFound,
@@ -24,7 +24,8 @@ pub fn App() -> impl IntoView {
                             <Route path="/pipelines/info" view=PipelineInfo />
                             <Route path="/pipelines/run" view=RunPipeline />
                             <Route path="/cron" view=CronJobs />
-                            <Route path="/cron/edit" view=CronJobsEdit />
+                            <Route path="/cron/insert" view=CronJobInsert />
+                            <Route path="/cron/update" view=CronJobUpdate />
                             <Route path="/monit" view=Monit />
                         </Route>
                         <Route path="/login" view=Login />
